Upgrade frontend local Docker base to node:16.20.2
This allows us to maintain package-lock.json with lockFileVersion 2.

Bump buildpack Node version to 16.

Updating to Node 16 requires using a fork of react-loading-overlay which supports React 17... which this app already uses, but it wont build in Node 16 with the unforked package. Not sure how this worked before, but this is what I've discovered.

Frontend wont build in Node > 16 so that what I've chosen. I tried Node 18 but it wont build, and complains of gyp errors.